[OPP-1422] bruke CheckboksPanelGruppe
Vi ønsker å bruke den innebygde CheckboksPanelGruppe-komponenten i stedet for en liste over checkbokser. På denne måten blir hele boksen klikkbar.
